+++ src/solution014.h
@@ -18,9 +18,9 @@ public:
         }
 
         std::string res = strs[0];
-        for (int i = 1; i < strs.size() - 1; i++) {
+        for (size_t i = 1; i < strs.size() - 1; i++) {
             std::string current_common;
-            for (int j = 0; j < strs[i].length() && j < strs[i+1].length() && strs[i][j] == strs[i+1][j]; j++) {
+            for (size_t j = 0; j < strs[i].length() && j < strs[i+1].length() && strs[i][j] == strs[i+1][j]; j++) {
                 current_common += strs[i][j];
             }
             if (res.length() > current_common.length()) {
@@ -36,7 +36,7 @@ public:
         }
 
         std::string current_common = strs[0];
-        for (int i = 1; i < strs.size() - 1; i++) {
+        for (size_t i = 1; i < strs.size() - 1; i++) {
             int length = std::min(current_common.length(), strs[i].length());
             int j = 0;
             for (j = 0; j < length; j++) {
